intelli-credit/requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
python-docx>=1.0.0
plotly>=5.18.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
google-search-results>=2.4.2
python-dotenv>=1.0.0
